+++ tests/conftest.py
@@ -130,7 +130,10 @@ GOOGLE_ANALYTICS_CODE = 'abc123'
 def db_tables_session_scope(db_engines):
     from cnxdb.contrib.pytest import db_tables
     with warnings.catch_warnings():
-        # Ignore SQLAlchemy SAWarning about unsupported reflection elements.
+        # FIXME: This suppresses the warning for now...
+        #        RemovedInPytest4Warning: Fixture "db_tables" called directly.
+        #        Fixtures are not meant to be called directly, are created
+        #        automatically when test functions request them as parameters.
         warnings.simplefilter('ignore')
         tables = db_tables(db_engines)
     return tables